Code cleanup, finish implementing dto into Todos

+++ src/main/java/com/roasteg/todo/service/TodoService.java
@@ -2,11 +2,13 @@ package com.roasteg.todo.service;
 
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Stream;
 
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.roasteg.todo.dto.TodoDto;
 import com.roasteg.todo.exception.TodoNotFoundException;
+
 import org.springframework.stereotype.Service;
 
 import com.roasteg.todo.model.Todo;
@@ -30,30 +32,49 @@ public class TodoService {
     }
 
     public TodoDto saveTodo(TodoDto todo, String token) {
-        final Todo fromDto = mapper.convertValue(todo, Todo.class);
+        final Todo fromDto = toEntity(todo);
         fromDto.setUser(jwtService.getUserFromToken(token));
         final Todo savedTodo = todoRepository.save(fromDto);
-        return mapper.convertValue(savedTodo, TodoDto.class);
+        return toDto(savedTodo);
     }
 
     public TodoDto editTodo(int id, Map<String, Object> fields, String token) throws JsonMappingException {
-        final Todo todoToEdit = todoRepository.findById(id).orElseThrow(TodoNotFoundException::new);
-        if(todoToEdit.getUser() != jwtService.getUserFromToken(token)) {
+        final Todo todoToEdit = todoRepository.findByUser(jwtService.getUserFromToken(token))
+                .stream()
+                .filter(todo -> todo.getId() == id)
+                .findFirst()
+                .orElseThrow(TodoNotFoundException::new);
+        final Todo editedTodo = mapper.updateValue(todoToEdit, fields);
+        final Todo savedTodo = todoRepository.save(editedTodo);
+        return toDto(savedTodo);
+    }
+
+    public TodoDto changeTodoStatus(int id, String token) {
+        final Todo todoToFinish = todoRepository.findByUser(jwtService.getUserFromToken(token))
+                .stream()
+                .filter(todo -> todo.getId() == id)
+                .findFirst()
+                .orElseThrow(TodoNotFoundException::new);
+        todoToFinish.setFinished(!todoToFinish.getFinished());
+        final Todo savedTodo = todoRepository.save(todoToFinish);
+        return toDto(savedTodo);
+    }
+
+    public void deleteTodo(int id, String token) {
+        final Stream<Todo> userTodos = todoRepository.findByUser(jwtService.getUserFromToken(token)).stream();
+        if(userTodos.anyMatch(todo -> todo.getId() == id)) {
+            todoRepository.deleteById(id);
+        } else {
             throw new TodoNotFoundException();
         }
-        final Todo todo = mapper.updateValue(todoToEdit, fields);
-        final Todo savedTodo = todoRepository.save(todo);
-        return mapper.convertValue(savedTodo, TodoDto.class);
     }
 
-    public Todo changeTodoStatus(int id) {
-        final Todo todo = todoRepository.findById(id).orElseThrow(TodoNotFoundException::new);
-        todo.setFinished(!todo.getFinished());
-        return todoRepository.save(todo);
+    private TodoDto toDto(Todo todo) {
+        return mapper.convertValue(todo, TodoDto.class);
     }
 
-    public void deleteTodo(int id) {
-        todoRepository.deleteById(id);
-    }
+    private Todo toEntity(TodoDto todoDto) {
+        return mapper.convertValue(todoDto, Todo.class);
+    } 
 
 }
